% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ascend.R
\name{apply_ascend}
\alias{apply_ascend}
\title{Apply ascend}
\usage{
apply_ascend(sce, params, k)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{params}{A list of parameter settings}

\item{k}{An integer giving the number of clusters to generate}
}
\value{
A list with timing information, cluster assignments and the estimated
  number of clusters.
}
\description{
Apply ascend
}
