% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_SAFE.R
\name{apply_SAFE}
\alias{apply_SAFE}
\title{Apply SAFE clustering}
\usage{
apply_SAFE(sce, params, k)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{params}{A list of parameter settings}

\item{k}{An integer giving the number of clusters to generate}
}
\value{
A list with timing information, cluster assignments and the estimated
  number of clusters.
}
\description{
Apply SAFE clustering
}
