% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_SC3svm.R
\name{apply_SC3svm}
\alias{apply_SC3svm}
\title{Apply SC3svm}
\usage{
apply_SC3svm(sce, params, k)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{params}{A list of parameter settings}

\item{k}{An integer giving the number of clusters to generate}
}
\value{
A list with timing information, cluster assignments and the estimated
  number of clusters.
}
\description{
Apply SC3svm
}
