% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Seurat.R
\name{apply_Seurat}
\alias{apply_Seurat}
\title{Apply Seurat}
\usage{
apply_Seurat(sce, params, resolution)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{params}{A list of parameter settings}

\item{resolution}{An numeric value giving the resolution to use for the clustering}
}
\value{
A list with timing information, cluster assignments and the estimated
  number of clusters.
}
\description{
Apply Seurat
}
