% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stability.R
\name{ARIdf}
\alias{ARIdf}
\title{Help function for computing ARI}
\usage{
ARIdf(x)
}
\arguments{
\item{x}{A \code{data.frame} with clustering results.}
}
\value{
a \code{data.frame} with ARI values for each pair of runs.
}
\description{
Help function for computing ARI
}
