\name{duo_clustering_all_parameter_settings_v1}
\alias{duo_clustering_all_parameter_settings_v1}

\title{
Hyperparameter values
}
\arguments{
\item{metadata}{Logical, whether only metadata should be returned}
}
\description{
Hyperparameter values for all clustering algorithms and data sets in v1 of Duo et al (F1000Research 2018)
}
\details{
List of hyperparametervalues used used for all clustering algorithms and data sets in v1 of Duo et al (F1000Research 2018).
}
\usage{
duo_clustering_all_parameter_settings_v1(metadata = FALSE)
}
\examples{
duo_clustering_all_parameter_settings_v1()
}
\value{Returns a \code{\link{list}}.}
\references{
Duò, A., Robinson, M.D., and Soneson, C. (2018).
\emph{A systematic performance evaluation of clustering methods for single-cell RNA-seq data.}
F1000Research, 7:1141.
.
}
\keyword{datasets}

