\name{duo_clustering_all_parameter_settings_v2}
\alias{duo_clustering_all_parameter_settings_v2}

\title{
Hyperparameter values
}
\arguments{
\item{metadata}{Logical, whether only metadata should be returned}
}
\description{
Hyperparameter values for all clustering algorithms and data sets in
v2 of Duo et al (F1000Research 2018)
}
\details{
List of hyperparameter values used for all clustering algorithms and
data sets in v2 of Duò et al (F1000Research 2018).
}
\usage{
duo_clustering_all_parameter_settings_v2(metadata = FALSE)
}
\examples{
duo_clustering_all_parameter_settings_v2()
}
\value{Returns a \code{list} with hyperparameter values for all data sets and
methods.}
\references{
Duò, A., Robinson, M.D., and Soneson, C. (2018).
\emph{A systematic performance evaluation of clustering methods for
single-cell RNA-seq data.}
F1000Research, 7:1141.
}
\keyword{datasets}

