\name{DuoClustering2018}
\docType{package}
\alias{DuoClustering2018}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
DuoClustering2018
}
\description{
Data package containing scRNA-seq data sets, clustering results and functions
for running and summarizing the performance of different scRNA-seq clustering
methods.
}

\details{
This package contains publicly available scRNA-seq data sets and the
accompanying results from clustering using general-purpose methods and scRNA-seq
clustering methods. Several real data sets a well as simulated data sets are
provided. The data sets have been used to evaluate the performance of clustering
algorithms in our previous work and publication (Duò et al.,
F1000Research 2018). The data sets are available as
\code{SingleCellExperiment} objects. For additional details on the data
sets, see the help files for the respective data sets.

Additionally, the clustering results from the evaluation as well as functions
for running the clustering methods and functions for summarization and
visualization of the clustering results are provided.

A description of the basic usage of the package for retrieving data sets and
clustering results, and how to construct various plots summarizing the
performance of different methods is outlined in the package vignettes.
}
\author{
Angelo Duò and Charlotte Sonesson
}
\references{
Duò, A., Robinson, M.D., and Soneson, C. (2018).
\emph{A systematic performance evaluation of clustering methods for
single-cell RNA-seq data.}
F1000Research, 7:1141.
}

