% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_clustering_modified.R
\name{individual_clustering}
\alias{individual_clustering}
\title{Single-cell Aggregated clustering From Ensemble (SAFE)}
\usage{
individual_clustering(inputTags, datatype = "count", SC3 = TRUE,
  gene_filter = FALSE, svm_num_cells = 5000, CIDR = TRUE,
  Seurat = TRUE, nPC = NULL, resolution = 0.9,
  seurat_min_cell = 200, resolution_min = 1.2, tSNE = TRUE,
  var_genes = NULL, SEED = 1, n_cores)
}
\arguments{
\item{inputTags}{a G*N matrix with G genes and N cells.}

\item{datatype}{defines the type of data, which could be "count", "CPM", "RPKM" and "FPKM".
Default is "count".}

\item{SC3}{a boolean variable that defines whether to cluster cells using SC3 method.
Default is "TRUE".}

\item{gene_filter}{a boolean variable defines whether to perform gene filtering
before SC3 clustering, when \code{SC3 = TRUE}.}

\item{CIDR}{a boolean parameter that defines whether to cluster cells using CIDR method.
Default is "TRUE".}

\item{Seurat}{is a boolean variable that defines whether to cluster cells using Seurat method.
Default is "TRUE".}

\item{nPC}{defines the number of princple compoents used in Seurat clustering, when \code{Seurat = TRUE}.
Default value is esimated by \code{nPC} of \code{CIDR}.}

\item{resolution}{defines the value of resolution used in Seurat clustering, when \code{Seurat = TRUE}.}

\item{seurat_min_cell}{defines the mimimum number of cells in input dataset below which
\code{resolution} is set to 1.2, when \code{Seurat = TRUE}.}

\item{resolution_min}{defines the resolution used in Seurat clustering for small dataset,
when \code{Seurat == TRUE} and cell number of input file < \code{seurat_min_cell}.}

\item{tSNE}{is a boolean variable that defines whether to cluster cells using tSNE method.
Default is "TRUE".}

\item{var_genes}{defines the number of variable genes used by tSNE analysis, when \code{tSNE = TRUE}.}

\item{SEED}{sets the seed of the random number generator. Setting the seed to a fixed value can
produce reproducible clustering results.}

\item{svm_max, }{if \code{SC3 = TRUE}, then defines the mimimum number of cells above which SVM will be run.}
}
\value{
a matrix of indiviudal clustering results is returned.
}
\description{
This function performs single-cell clustering using four state-of-the-art methods,
SC3, CIDR, Seurat and tSNE+kmeans.
}
\references{
Yuchen Yang, Ruth Huh, Houston Culpepper, Yun Li. SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data. 2017
}
\author{
Yuchen Yang <yyuchen@email.unc.edu>, Ruth Huh <rhuh@live.unc.edu>,
Houston Culpepper <hculpepp@live.unc.edu>, Yun Li <yunli@med.unc.edu>
}
