% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Plot entropy of cluster assignments}
\usage{
plot_entropy(res, method_colors = NULL)
}
\arguments{
\item{res}{A data.frame with clustering results.}

\item{method_colors}{A named vector with colors to use for the different
clustering methods. Can be NULL, in which case colors are chosen
automatically.}
}
\value{
A named list of ggplot2 objects
}
\description{
Plot entropy of cluster assignments
}
\examples{
res <- clustering_summary_filteredExpr10_Koh_v2()
plots <- plot_entropy(res)

}
\author{
Angelo Duo, Charlotte Soneson
}
