% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_diff.R
\name{plot_k_diff}
\alias{plot_k_diff}
\title{Plot differences between optimal, estimated and true number of clusters}
\usage{
plot_k_diff(res, method_colors = NULL)
}
\arguments{
\item{res}{A data.frame with clustering results.}

\item{method_colors}{A named vector with colors to use for the different
clustering methods. Can be NULL, in which case colors are chosen
automatically.}
}
\value{
A named list of ggplot2 objects
}
\description{
Plot differences between optimal, estimated and true number of clusters
}
\examples{

}
\author{
Angelo Duo, Charlotte Soneson
}
