% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stability.R
\name{plot_stability}
\alias{plot_stability}
\title{Plot stability of methods}
\usage{
plot_stability(res, method_colors = NULL)
}
\arguments{
\item{res}{A data.frame with clustering results.}

\item{method_colors}{A named vector with colors to use for the different
clustering methods. Can be NULL, in which case colors are chosen
automatically.}
}
\value{
A named list of \code{ggplot2} objects
}
\description{
Plot the stability of the clusterings obtained for each method
}
\examples{
res <- clustering_summary_filteredExpr10_Koh_v1()
plots <- plot_stability(res)

}
\author{
Angelo Duo, Charlotte Soneson
}
