% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE_modified.R
\name{SAFE}
\alias{SAFE}
\title{Single-cell Aggregated clustering From Ensemble (SAFE)}
\usage{
SAFE(cluster_results, k_min = 2, k_max = NULL, MCLA = TRUE,
  HGPA = FALSE, CSPA = FALSE, cspc_cell_max = NULL, SEED = 1)
}
\arguments{
\item{cluster_results}{a J*N matrix with J individual clustering methods and N cells}

\item{k_min}{defines the minimum number of clusters used for ensembel clustering. Default is 2.}

\item{k_max}{defines the maximum number of clusters used for ensembel clustering. Default is
the maximum cluster number estimated by all the single solutions.}

\item{MCLA}{a boolean parameter that defines whether to use MCLA algorithm for ensemble clustering.
Default is "TRUE".}

\item{HGPA}{a boolean parameter that defines whether to use HGPA algorithm for ensemble clustering.
Default is "FALSE".}

\item{CSPA}{a boolean parameter that defines whether to use CSPA algorithm for ensemble clustering.
Default is "FALSE".}

\item{cspc_cell_max}{defines the maximum number of cells above which CSPA is not run, when \code{CSPA = TRUE}.}

\item{SEED}{sets the seed of the random number generator. Setting the seed to a fixed value can produce
reproducible clustering results for MCLA and CSPA algorithms.}
}
\value{
a list containing the overall optimal ensemble clustering and cluster number,
as well as the optimal resutls for each algorithm.
}
\description{
SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data
}
\references{
Yuchen Yang, Ruth Huh, Houston Culpepper, Yun Li. SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data. 2017
}
\author{
Yuchen Yang <yyuchen@email.unc.edu>, Ruth Huh <rhuh@live.unc.edu>,
Houston Culpepper <hculpepp@live.unc.edu>, Yun Li <yunli@med.unc.edu>
}
