% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Calculate Shannon entropy}
\usage{
shannon_entropy(cluster_assignments)
}
\arguments{
\item{cluster_assignments}{A vector with cluster assignments}
}
\value{
The Shannon entropy of the assignment vector
}
\description{
Calculate Shannon entropy
}
