% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy.R
\name{shannonEntropy}
\alias{shannonEntropy}
\title{Calculate Shannon entropy}
\usage{
shannonEntropy(clusterAssignments)
}
\arguments{
\item{clusterAssignments}{A vector with cluster assignments}
}
\value{
The Shannon entropy of the assignment vector
}
\description{
Calculate Shannon entropy
}
